Fix integration test for AR screen
It smeens to be fixed for the moment, tested locally on device with and without AR Core
